+++ ui/packages/shared/store/src/index.tsx
@@ -15,6 +15,6 @@ import store from './store';
 import {useAppDispatch, useAppSelector} from './hooks';
 export * from './slices/uiSlice';
 export * from './slices/profileSlice';
-export * from './store';
+export * from './slices/colorsSlice';
 
 export {store, useAppDispatch, useAppSelector};