+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.104](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.103...@parca/profile@0.16.104) (2023-01-27)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.103](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.102...@parca/profile@0.16.103) (2023-01-27)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.102](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.101...@parca/profile@0.16.102) (2023-01-26)
+
+### Bug Fixes
+
+- change Reset zoom button text to be Reset view ([0061637](https://github.com/parca-dev/parca/commit/0061637873ca5dc7633a8ee9df7883779cd469d1))
+
+## [0.16.101](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.96...@parca/profile@0.16.101) (2023-01-25)
+
+**Note:** Version bump only for package @parca/profile
+
 ## 0.16.100 (2023-01-22)
 
 **Note:** Version bump only for package @parca/profile